Unpin zeroize dependency
The pin of zeroize `<1.4` caused issues with my dependency list. I
understand the interest to maintain the MSRV, but I think MSRV should be
guaranteed by adding a `Cargo.lock` to crates that use this crate. In my
opinion, constraining the upper-bound only makes sense if the crate to
depend on (in this case `zeroize`) somehow violated semantic versioning
and introduced a breaking change in a minor version. This is (as far as
I have seen) not the case. Thus, I would suggest removing the
upper-bound; the MSRV still holds as one can choose lower versions of
zeroize in the `Cargo.toml` or `Cargo.lock` in crates using this version
of `curve25519-dalek`.